+++ NEWS.md
@@ -5,6 +5,7 @@
 ## New features
 
 * Added interface for the Bayesian Structural Time Series model (via the 'bsts' R package) to the fable framework.
+* Added static intercepts, autoregressive, level and linear trends, and multiple regression seasonalities.
 * Added bsts model methods for: `forecast()`, `components()`, `fitted()`, `residuals()`.
 * Added package introduction vignette.